backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
cloudinary==1.41.0
python-multipart==0.0.12
Pillow==10.4.0
pillow-heif==0.18.0
google-genai
google-cloud-aiplatform>=1.70.0
pytest
pytest-asyncio
